feat: Use crane for image loader instead of skopeo
